# avoid clearing fast node cache during pruning